Ensure WASM tests can run
